Safari-609.1.11.1/LayoutTests/http/tests/ssl/media-stream/get-user-media-different-host.html

<!DOCTYPE HTML>
<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
        <script src="resources/get-user-media.js"></script>        
        <script>

            let testInfo = { 
                shouldSucceed : false,
                depth: 1,
                mainFrame : { protocol : "https:", host: "127.0.0.1", port: 8443 },
                iFrame : { protocol : "https:", host: "localhost", port: 8443, pathname: "/ssl/media-stream/resources/get-user-media-frame.html" }
            };

            description("Tests that getUserMedia fails when the top level document and iframe do not have the same domain.");

        </script>
    </head>
    <body>
        <div id="console"></div>
        <script src="../../../../resources/js-test-post.js"></script>
    </body>
</html>
